requirements.txt
Django>=4.2,<5.0
openpyxl>=3.1.2
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.126.0
